feat(customer): add customer availability, booking, location, product, and schedule functions
This commit adds new files and functions related to customer availability, booking, location, product, and schedule. It includes the following changes:

- Added customer-availability.function.ts file with the CustomerAvailabilityControllerGet function.
- Added customer-booking.function.ts file with the CustomerBookingControllerGet and CustomerBookingControllerList functions.
- Added customer-location.function.ts file with the CustomerLocationControllerAdd, CustomerLocationControllerCreate, CustomerLocationControllerGetAll, CustomerLocationControllerGetAllOrigins, CustomerLocationControllerGetOne, CustomerLocationControllerRemove, CustomerLocationControllerSetDefault, and CustomerLocationControllerUpdate functions.
- Added customer-product.function.ts file with the CustomerProductsControllerList, CustomerProductControllerDestroy, CustomerProductControllerGet, and CustomerProductControllerUpsert functions.
- Added customer-schedule.function.ts file with the CustomerScheduleControllerCreate, CustomerScheduleControllerDestroy, CustomerScheduleControllerGet, Customer

chore(get.spec.ts): update import path for CustomerScheduleServiceCreate in get.spec.ts
chore(upsert.spec.ts): update import path for CustomerScheduleServiceCreate in upsert.spec.ts
chore(list-ids.spec.ts): update import path for CustomerScheduleServiceCreate in list-ids.spec.ts
chore(list.spec.ts): update import path for CustomerScheduleServiceCreate in list.spec.ts
chore(create.ts): update import path for CustomerScheduleServiceCreate in create.ts
chore(destroy.spec.ts): update import path for CustomerScheduleServiceCreate in destroy.spec.ts
chore(destroy.ts): update import path for CustomerScheduleServiceDestroy in destroy.ts
chore(get.spec.ts): update import path for CustomerScheduleServiceCreate in get.spec.ts
chore(get.ts): update import path for CustomerScheduleServiceGet in get.ts
chore(list.spec.ts): update import path for CustomerScheduleServiceList in list.spec.ts
chore(list.ts

feat(schedule): add create, destroy, get, and getWithCustomer services

feat(schedule/create): add CustomerScheduleServiceCreate function

feat(schedule/destroy): add CustomerScheduleServiceDestroy function

feat(schedule/get): add CustomerScheduleServiceGet function

feat(schedule/getWithCustomer): add CustomerScheduleServiceGetWithCustomer function

test(schedule/create.spec): add tests for CustomerScheduleServiceCreate function

test(schedule/destroy.spec): add tests for CustomerScheduleServiceDestroy function

test(schedule/get.spec): add tests for CustomerScheduleServiceGet function

test(schedule/get-with-customer.spec): add tests for CustomerScheduleServiceGetWithCustomer function

feat(schedule): add list, update, and get-with-customer services

feat(schedule/list): add service to retrieve all schedules for a customerId

feat(schedule/update): add service to update a schedule

feat(schedule/get-with-customer): add service to retrieve a schedule with customer information

fix(schedule): fix import path in user/services/schedule.spec.ts

fix(availability): fix import path in generate-availability.ts
