+++ src/functions/customer.function.ts
@@ -1,13 +1,7 @@
 import "module-alias/register";
-import { CustomerProductsControllerList } from "./customer/controllers/products/list";
 
 import { app } from "@azure/functions";
 
-import { CustomerAvailabilityControllerGet } from "./customer/controllers/availability";
-import {
-  CustomerBookingControllerGet,
-  CustomerBookingControllerList,
-} from "./customer/controllers/booking";
 import {
   CustomerControllerGet,
   CustomerControllerIsBusiness,
@@ -15,30 +9,6 @@ import {
   CustomerControllerUpdate,
   CustomerControllerUpsert,
 } from "./customer/controllers/customer";
-import {
-  CustomerLocationControllerAdd,
-  CustomerLocationControllerCreate,
-  CustomerLocationControllerGetAll,
-  CustomerLocationControllerGetAllOrigins,
-  CustomerLocationControllerGetOne,
-  CustomerLocationControllerRemove,
-  CustomerLocationControllerSetDefault,
-  CustomerLocationControllerUpdate,
-} from "./customer/controllers/location";
-import {
-  CustomerProductControllerDestroy,
-  CustomerProductControllerGet,
-  CustomerProductControllerUpsert,
-} from "./customer/controllers/product";
-import { CustomerProductsControllerListIds } from "./customer/controllers/products";
-import {
-  CustomerScheduleControllerCreate,
-  CustomerScheduleControllerDestroy,
-  CustomerScheduleControllerGet,
-  CustomerScheduleControllerList,
-  CustomerScheduleControllerUpdate,
-} from "./customer/controllers/schedule";
-import { CustomerScheduleSlotControllerUpdate } from "./customer/controllers/slot";
 
 /* Customer */
 
@@ -75,174 +45,4 @@ app.http("customerStatus", {
   authLevel: "anonymous",
   route: "customer/{customerId?}/status",
   handler: CustomerControllerStatus,
-});
-
-/* **************** */
-/* Customer/Product */
-/* **************** */
-
-app.http("customerProductsListIds", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/products/ids",
-  handler: CustomerProductsControllerListIds,
-});
-
-app.http("customerProductsList", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/products",
-  handler: CustomerProductsControllerList,
-});
-
-app.http("customerAvailabilityGet", {
-  methods: ["POST"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/availability/{locationId?}/get",
-  handler: CustomerAvailabilityControllerGet,
-});
-
-app.http("customerProductUpsert", {
-  methods: ["PUT"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/product/{productId?}",
-  handler: CustomerProductControllerUpsert,
-});
-
-app.http("customerProductGet", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/product/{productId?}",
-  handler: CustomerProductControllerGet,
-});
-
-app.http("customerProductDestroy", {
-  methods: ["DELETE"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/product/{productId?}",
-  handler: CustomerProductControllerDestroy,
-});
-
-/* ******* */
-/* Booking */
-/* ******* */
-
-app.http("customerBookingGet", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/booking/{orderId?}",
-  handler: CustomerBookingControllerGet,
-});
-
-app.http("customerBookingList", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/bookings",
-  handler: CustomerBookingControllerList,
-});
-
-/* ******* */
-/* Location */
-/* ******* */
-
-app.http("customerLocationGetAllOrigins", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/locations/get-all-origins",
-  handler: CustomerLocationControllerGetAllOrigins,
-});
-
-app.http("customerLocationList", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/locations",
-  handler: CustomerLocationControllerGetAll,
-});
-
-app.http("customerLocationGet", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/location/{locationId}",
-  handler: CustomerLocationControllerGetOne,
-});
-
-app.http("customerLocationUpdate", {
-  methods: ["PUT"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/location/{locationId?}",
-  handler: CustomerLocationControllerUpdate,
-});
-
-app.http("customerLocationCreate", {
-  methods: ["POST"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/locations",
-  handler: CustomerLocationControllerCreate,
-});
-
-app.http("customerLocationAdd", {
-  methods: ["POST"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/location/{locationId}",
-  handler: CustomerLocationControllerAdd,
-});
-
-app.http("customerLocationRemove", {
-  methods: ["DELETE"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/location/{locationId}",
-  handler: CustomerLocationControllerRemove,
-});
-
-app.http("customerLocationSetDefault", {
-  methods: ["PUT"],
-  authLevel: "anonymous",
-  route: "customer/{customerId}/location/{locationId}/setDefault",
-  handler: CustomerLocationControllerSetDefault,
-});
-
-/* ******** */
-/* Schedule
-/* ******** */
-
-app.http("customerScheduleDestroy", {
-  methods: ["DELETE"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedule/{scheduleId?}",
-  handler: CustomerScheduleControllerDestroy,
-});
-
-app.http("customerScheduleCreate", {
-  methods: ["POST"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedule",
-  handler: CustomerScheduleControllerCreate,
-});
-
-app.http("customerScheduleUpdate", {
-  methods: ["PUT"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedule/{scheduleId?}",
-  handler: CustomerScheduleControllerUpdate,
-});
-
-app.http("customerScheduleGet", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedule/{scheduleId?}",
-  handler: CustomerScheduleControllerGet,
-});
-
-app.http("customerScheduleList", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedules",
-  handler: CustomerScheduleControllerList,
-});
-
-app.http("customerScheduleSlotUpdate", {
-  methods: ["PUT"],
-  authLevel: "anonymous",
-  route: "customer/{customerId?}/schedule/{scheduleId?}/slots",
-  handler: CustomerScheduleSlotControllerUpdate,
 });
