+++ src/functions/customer/controllers/product/get.spec.ts
@@ -8,7 +8,7 @@ import {
   createHttpRequest,
 } from "~/library/jest/azure";
 import { CustomerProductServiceUpsert } from "../../services/product";
-import { CustomerScheduleServiceCreate } from "../../services/schedule";
+import { CustomerScheduleServiceCreate } from "../../services/schedule/create";
 import {
   CustomerProductControllerGet,
   CustomerProductControllerGetRequest,
