+++ src/functions/customer/controllers/products/list-ids.spec.ts
@@ -7,7 +7,7 @@ import {
   createContext,
   createHttpRequest,
 } from "~/library/jest/azure";
-import { CustomerScheduleServiceCreate } from "../../services/schedule";
+import { CustomerScheduleServiceCreate } from "../../services/schedule/create";
 import {
   CustomerProductsControllerListIds,
   CustomerProductsControllerListIdsRequest,
