+++ src/functions/customer/controllers/schedule/destroy.spec.ts
@@ -4,7 +4,7 @@ import {
   createContext,
   createHttpRequest,
 } from "~/library/jest/azure";
-import { CustomerScheduleServiceCreate } from "../../services/schedule";
+import { CustomerScheduleServiceCreate } from "../../services/schedule/create";
 import {
   CustomerScheduleControllerDestroy,
   CustomerScheduleControllerDestroyRequest,
