+++ src/functions/customer/controllers/schedule/list.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 import { ScheduleZodSchema } from "~/functions/schedule/schedule.types";
 import { _ } from "~/library/handler";
-import { CustomerScheduleServiceList } from "../../services/schedule";
+import { CustomerScheduleServiceList } from "../../services/schedule/list";
 
 export type CustomerScheduleControllerListRequest = {
   query: z.infer<typeof CustomerScheduleControllerListQuerySchema>;
