+++ src/functions/customer/services/schedule/destroy.ts
@@ -1,0 +1,15 @@
+import { Schedule, ScheduleModel } from "~/functions/schedule";
+
+export type CustomerScheduleServiceDestroyProps = {
+  scheduleId: Schedule["_id"];
+  customerId: Schedule["customerId"];
+};
+
+export const CustomerScheduleServiceDestroy = async (
+  props: CustomerScheduleServiceDestroyProps
+) => {
+  return ScheduleModel.deleteOne({
+    _id: props.scheduleId,
+    customerId: props.customerId,
+  });
+};
