+++ src/library/availability/generate-availability.ts
@@ -9,14 +9,15 @@ import {
 import { utcToZonedTime } from "date-fns-tz";
 import { enUS } from "date-fns/locale";
 import { Availability } from "~/functions/availability";
-import { CustomerScheduleServiceGetWithCustomerResponse } from "~/functions/customer/services/schedule";
+
 import { LookupServiceCreate } from "~/functions/lookup";
 import { calculateMaxNoticeAndMinBookingPeriod } from "./calculate-max-notice-and-min-booking-period";
 
+import { CustomerScheduleServiceGetWithCustomer } from "~/functions/customer/services/schedule/get-with-customer";
 import { generateEndDate, generateStartDate } from "./start-end-date";
 
 export type GenerateAvailabilityProps = {
-  schedule: CustomerScheduleServiceGetWithCustomerResponse;
+  schedule: Awaited<ReturnType<typeof CustomerScheduleServiceGetWithCustomer>>;
   lookup?: Awaited<ReturnType<typeof LookupServiceCreate>>;
   startDate: string;
 };
